feat(Landmarks): patch to convert image landmarks to the new structure.
Added semantic patch to convert the "m_imageLandmarksId" field of image from
PointList to Landmarks.
Added graphlink from V09ALA to V10.
Now the data version is V10.

+++ SrcLib/core/fwAtomConversion/src/fwAtomConversion/mapper/Landmarks.cpp
@@ -175,8 +175,8 @@ fwAtomConversionRegisterMacro( ::fwAtomConversion::mapper::Landmarks, ::fwData::
                                    resultPt.size() != 3 );
 
             ::fwData::Landmarks::PointType pt = {{
-                                                     std::stof(resultPt[0]), std::stof(resultPt[1]),
-                                                     std::stof(resultPt[2])
+                                                     std::stod(resultPt[0]), std::stod(resultPt[1]),
+                                                     std::stod(resultPt[2])
                                                  }};
             landmarks->addPoint(name, pt);
         }
